refactor(agent): split stream execution modules
Replace the stream execution helper monolith with focused state, event, result, policy, turn, flow, prompt, completion, and writeback modules.

Keep stream engine behavior intact while routing internal callers and tests to concrete modules.
